it/python/pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "integration-test"
authors = [{ name = "Microsoft", email = "graphtooling+python@microsoft.com" }]
dependencies = [
    "microsoft-kiota-abstractions >= 1.0.0",
    "microsoft-kiota-http >= 1.2.0",
    "microsoft-kiota-authentication-azure >= 1.0.0",
    "microsoft-kiota-serialization-json >= 1.0.0",
    "microsoft-kiota-serialization-text >= 1.0.0",
]
license = { file = "LICENSE" }
readme = "README.md"
keywords = ["kiota"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
]
dynamic = ["version", "description"]


[project.urls]
homepage = "https://github.com/microsoft/kiota#readme"
repository = "https://github.com/microsoft/kiota"
documentation = "https://learn.microsoft.com/openapi/kiota"

[tool.flit.module]
name = "integration_test"

[mypy]
warn_unused_configs = true
files = "integration_test"
ignore_missing_imports = true

[tool.yapf]
based_on_style = "pep8"
dedent_closing_brackets = true
each_dict_entry_on_separate_line = true
column_limit = 100

[tool.isort]
profile = "hug"
